airplane-flight-tracker-server/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
Flask==2.3.3
Flask-Cors==4.0.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
typing_extensions==4.8.0
Werkzeug==2.3.7
